[APPS] Regenrated fs_qspi_flash_bootloader for pic32cz_ca80_cult.X
MH3-110298

+++ apps/fs/qspi/app_programmer/firmware/src/config/pic32cz_ca80_cult/driver/sdspi/src/drv_sdspi.c
@@ -2529,7 +2529,7 @@ static void DRV_SDSPI_SetupXfer(
         {
             return;
         }
-        if (((blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_READ_ENTRY].numBlocks))
+        if ((((uint64_t)blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_READ_ENTRY].numBlocks))
         {
             return;
         }
@@ -2540,7 +2540,7 @@ static void DRV_SDSPI_SetupXfer(
         {
             return;
         }
-        if (((blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_WRITE_ENTRY].numBlocks))
+        if ((((uint64_t)blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_WRITE_ENTRY].numBlocks))
         {
             return;
         }
